telemetry: use _EXPERIMENTAL env variables
