+++ cmd/upload-journal/main.go
@@ -7,9 +7,12 @@ import (
 	"os"
 	"os/signal"
 	"syscall"
+	"time"
 
 	"github.com/dagger/dagger/core/pipeline"
 	"github.com/dagger/dagger/telemetry"
+	bkclient "github.com/moby/buildkit/client"
+	"github.com/nxadm/tail"
 )
 
 func main() {
@@ -93,4 +96,50 @@ func processJournal(t *telemetry.Telemetry, entries chan *JournalEntry) error {
 	}
 
 	return nil
+}
+
+type JournalEntry struct {
+	Event *bkclient.SolveStatus
+	TS    time.Time
+}
+
+func tailJournal(journal string, follow bool, stopCh chan struct{}) (chan *JournalEntry, error) {
+	f, err := tail.TailFile(journal, tail.Config{Follow: follow})
+	if err != nil {
+		return nil, err
+	}
+
+	ch := make(chan *JournalEntry)
+
+	go func() {
+		if stopCh == nil {
+			return
+		}
+		<-stopCh
+		fmt.Fprintf(os.Stderr, "quitting\n")
+		if err := f.StopAtEOF(); err != nil {
+			fmt.Fprintf(os.Stderr, "err: %v\n", err)
+		}
+	}()
+
+	go func() {
+		defer close(ch)
+		defer f.Cleanup()
+
+		for line := range f.Lines {
+			if err := line.Err; err != nil {
+				fmt.Fprintf(os.Stderr, "err: %v\n", err)
+				return
+			}
+			var entry JournalEntry
+			if err := json.Unmarshal([]byte(line.Text), &entry); err != nil {
+				fmt.Fprintf(os.Stderr, "err: %v\n", err)
+				return
+			}
+
+			ch <- &entry
+		}
+	}()
+
+	return ch, nil
 }
